+++ recordreplay-servlet/src/main/java/be/xplore/recordreplay/servlet/HttpServletResponseMapper.java
@@ -1,4 +1,4 @@
-package be.xplore.recordreplay.service;
+package be.xplore.recordreplay.servlet;
 
 import be.xplore.fakes.model.Response;
